+++ src/main/java/com/example/whiplash/article/entity/UserArticleAssignment.java
@@ -1,7 +1,7 @@
 package com.example.whiplash.article.entity;
 
 import com.example.whiplash.domain.entity.history.email.EmailSendStatus;
-import com.example.whiplash.user.User;
+import com.example.whiplash.user.domain.User;
 import lombok.*;
 import jakarta.persistence.*;
 import java.time.LocalDateTime;
