+++ src/main/java/com/example/whiplash/article/repository/UserArticleAssignmentRepository.java
@@ -2,7 +2,7 @@ package com.example.whiplash.article.repository;
 
 import com.example.whiplash.article.entity.UserArticleAssignment;
 import com.example.whiplash.domain.entity.history.email.EmailSendStatus;
-import com.example.whiplash.user.User;
+import com.example.whiplash.user.domain.User;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.List;
