+++ src/main/java/com/example/whiplash/auth/service/AuthService.java
@@ -5,10 +5,13 @@ import com.example.whiplash.apiPayload.exception.WhiplashException;
 import com.example.whiplash.config.security.jwt.JwtTokenProvider;
 import com.example.whiplash.converter.AuthConverter;
 import com.example.whiplash.converter.UserConverter;
-import com.example.whiplash.user.User;
-import com.example.whiplash.user.UserStatus;
-import com.example.whiplash.user.dto.*;
+import com.example.whiplash.user.domain.User;
+import com.example.whiplash.user.domain.UserStatus;
 import com.example.whiplash.user.repository.UserRepository;
+import com.example.whiplash.user.web.dto.response.KakaoUserInfoResponseDTO;
+import com.example.whiplash.user.web.dto.request.LoginRequestDTO;
+import com.example.whiplash.user.web.dto.response.TokenResponseDTO;
+import com.example.whiplash.user.web.dto.request.UserCreateDTO;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.security.authentication.UsernamePasswordAuthenticationToken;
