+++ src/main/java/com/example/whiplash/delivery/email/SmtpEmailSender.java
@@ -3,8 +3,6 @@ package com.example.whiplash.delivery.email;
 
 import com.example.whiplash.article.document.SummarizedArticle;
 import com.example.whiplash.article.repository.SummarizedArticleRepository;
-import com.example.whiplash.user.User;
-import com.example.whiplash.user.repository.UserRepository;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.mail.SimpleMailMessage;
